example/requirements.txt
PyYAML==5.3.1
JSON-log-formatter==0.3.0
python-json-logger==0.1.11
structlog==20.1.0
colorama==0.4.3